Several changes for getting the bilingual version working, and associated functionality. i. BuildBilingualIndex now works! (According to tests on europarl english-french data.) ii. TermVectorsFromLucene edited to properly enable reuse of basicDocVectors.
This isn't perfect yet, by any means: it's probably inefficient in how much state is kept around when BuildBilingualIndex reuses TermVectorsFromLucene objects, and the interface to TermVectorsFromLucene could probably be improved. Since neither of these is a really basic class, I haven't thought too deeply about this - just don't write code that depends on these without rethinking and probably refactoring.

iii. Added a scripts directory, initially with the preprocessing script for the europarl chapter alignment,
iv. Added READMEs to the test and scripts directories.
v. Changed build.xml file so that tests aren't automatically compiled (this is so that casual users of the source won't have to depend on JUnit).



git-svn-id: http://semanticvectors.googlecode.com/svn/trunk@61 483481f0-a63c-0410-a8a8-396719eec1a4