Implemented TRRI as per the RRI paper. This required adding another constructor to TermVectorsFromLucene that generates elemental term vectors for future recycling. Also entropy weighting has been added to LuceneUtils, to facilitate log-entropy weighting in DocVectors and IncrementalDocVectors. TRRI can now be performed by calling BuildIndex with the command line flags: -initialtermvectors random -trainingcycles 2
This implementation doesn't scale as well as the one used in the RRI paper, as this one had an opposite number to IncrementalDocVectors named IncrementalTermVectors - combining these two allows one to reflect without ever keeping the document vectors in RAM, which is critical for large data sets on small machines. Hope to add this in the near future.


git-svn-id: http://semanticvectors.googlecode.com/svn/trunk@307 483481f0-a63c-0410-a8a8-396719eec1a4