I'm in pursuit of a malevolent bug which is causing sets of zero vectors to be recorded during TRRI with incremental indexing. These changes appear to fix it, but I haven't quite got to the root of why this works:
(1) normalize the document vectors before adding them to term vectors by default
- this is a reasonable thing to do, and if we keep them as is on disk we retain the possibility of incremental indexing

However.... what fixed the bug was:
(2) change the number of dimensions written to disk from the "dimension" parameter to the length of the vector from the vector store

Mysteriously, I've yet to find evidence of a case where the vector-to-be-written and the dimension variable didn't match one another.  However, the bug is reproducible by changing back "i < tmpVector.length" to "i < dimension", and disappears upon reversing this change. 

More from me as new clues are uncovered, but I thought it would be better to have a working implementation checked in for now.


git-svn-id: http://semanticvectors.googlecode.com/svn/trunk@465 483481f0-a63c-0410-a8a8-396719eec1a4